+++ src/PimEnterprise/Bundle/EnrichBundle/DependencyInjection/PimEnterpriseEnrichExtension.php
@@ -29,16 +29,17 @@ class PimEnterpriseEnrichExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader->load('controllers.yml');
+        $loader->load('datagrid_listeners.yml');
+        $loader->load('event_listeners.yml');
+        $loader->load('event_subscribers.yml');
         $loader->load('form_types.yml');
         $loader->load('mass_actions.yml');
         $loader->load('parameters.yml');
-        $loader->load('controllers.yml');
-        $loader->load('event_listeners.yml');
-        $loader->load('datagrid_listeners.yml');
-        $loader->load('event_subscribers.yml');
         $loader->load('twig.yml');
-        $loader->load('view_elements/product.yml');
+        $loader->load('updaters.yml');
         $loader->load('view_elements/attribute_group.yml');
         $loader->load('view_elements/category.yml');
+        $loader->load('view_elements/product.yml');
     }
 }